backend/requirements.txt
fastapi
uvicorn
minio
psycopg2-binary
python-dotenv
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
httpx
